# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A directory API allowing Nym nodes and clients to connect to each other.",
        "title": "Nym Directory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "https://github.com/nymtech/nym-validator/license"
        },
        "version": "0.9.0-dev"
    },
    "paths": {
...
        "/api/mixmining": {
            "post": {
                "description": "Nym network monitor sends packets through the system and checks if they make it. The network monitor then hits this method to report whether the node was up at a given time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lets the network monitor create a new uptime status for a mix",
                "operationId": "addMixStatus",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MixStatus"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/batch": {
            "post": {
                "description": "Nym network monitor sends packets through the system and checks if they make it. The network monitor then hits this method to report whether nodes were up at a given time.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lets the network monitor create a new uptime status for multiple mixes",
                "operationId": "batchCreateMixStatus",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BatchMixStatus"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/fullreport": {
            "get": {
                "description": "Provides summary uptime statistics for last 5 minutes, day, week, and month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Retrieves a summary report of historical mix status",
                "operationId": "batchGetMixStatusReport",
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/node/{pubkey}/history": {
            "get": {
                "description": "Lists all mixnode statuses for a given node pubkey",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lists mixnode activity",
                "operationId": "listMixStatuses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mixnode Pubkey",
                        "name": "pubkey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MixStatus"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/node/{pubkey}/report": {
            "get": {
                "description": "Provides summary uptime statistics for last 5 minutes, day, week, and month",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Retrieves a summary report of historical mix status",
                "operationId": "getMixStatusReport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mixnode Pubkey",
                        "name": "pubkey",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/register/gateway": {
            "post": {
                "description": "On Nym nodes startup they register their presence indicating they should be alive and get added to the set of active nodes in the topology.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lets a gateway tell the directory server it's coming online",
                "operationId": "registerGatewayPresence",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GatewayRegistrationInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/register/mix": {
            "post": {
                "description": "On Nym nodes startup they register their presence indicating they should be alive and get added to the set of active nodes in the topology.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lets a mixnode tell the directory server it's coming online",
                "operationId": "registerMixPresence",
                "parameters": [
                    {
                        "description": "object",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MixRegistrationInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/api/mixmining/register/{id}": {
            "delete": {
                "description": "Messages sent by a node on powering down to indicate it's going offline so that it should get removed from active topology.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Unregister presence of node.",
                "operationId": "unregisterPresence",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node Identity",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
        "/api/mixmining/topology": {
            "get": {
                "description": "On Nym nodes startup they register their presence indicating they should be alive. This method provides a list of nodes which have done so.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mixmining"
                ],
                "summary": "Lists Nym mixnodes and gateways on the network alongside their reputation.",
                "operationId": "getTopology",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Topology"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.BatchMixStatus": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MixStatus"
                    }
                }
            }
        },
        "models.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.GatewayRegistrationInfo": {
            "type": "object",
            "required": [
                "clientsHost",
                "identityKey",
                "mixHost",
                "sphinxKey",
                "version"
            ],
            "properties": {
                "clientsHost": {
                    "type": "string"
                },
                "identityKey": {
                    "type": "string"
                },
                "incentivesAddress": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "mixHost": {
                    "type": "string"
                },
                "sphinxKey": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "models.MixStatus": {
            "type": "object",
            "required": [
                "ipVersion",
                "pubKey",
                "up"
            ],
            "properties": {
                "ipVersion": {
                    "type": "string"
                },
                "pubKey": {
                    "type": "string"
                },
                "up": {
                    "type": "boolean"
                }
            }
        },
...
    }
}